Update ForEach activity with local variables
