+++ src/modules/Elsa.Workflows.Core/Activities/ParallelForEachT.cs
@@ -45,7 +45,7 @@ public class ParallelForEach<T> : Activity
         foreach (var item in items)
         {
             // For each item, declare a new variable the work to be scheduled.
-            var variable = new Variable<T>("CurrentItem", item)
+            var variable = new Variable<T>("CurrentValue", item)
             {
                 // TODO: This should be configurable, because this won't work for e.g. file streams and other non-serializable types.
                 StorageDriverType = typeof(WorkflowStorageDriver)
